# Speed Improvement through less looping